Presets Added, Promotion Section Underway
New feature for Presets have been added - an 80s Artists playlist and 90s Country Songs

Working on a promotional section to append to the bottom of the squares. Left side will be for a logo, right side will be for text promotion area. Considering adding an option to choose one or both, default is neither.